Notifications - Doctrine Table inheritance : https://www.udemy.com/learn-symfony-4-hands-on-creating-a-real-world-application/learn/lecture/10199206#overview
The goal is to Map different objects that inherit from each other, and have some common features.

With Single Table Inheritance, we would like an object (i.e. Notification.php) to be mapped to a single database table.

That's why each of the table's record will have all the fields from the based class, and from the classes that extends this based class.

So let's see what are the common things for a Notification.

For example, we have a property "user" in the "Notification" class, so all the child classes that extend the Notification entity will have a "user" id field.

To specify that a class could have inheritance classes, we use these annotations :
- "@ORM\InheritanceType("SINGLE_TABLE")"
- "@ORM\DiscriminatorColumn("discr", type="string")
- "@ORM\DiscriminatorMap({"like" = "LikeNotification"})