Notification Repository : Fetching Notifications useen by User. --------------------------------------------------------------- https://www.udemy.com/learn-symfony-4-hands-on-creating-a-real-world-application/learn/lecture/10199206#overview
NotificationRepository.php => add a query that would fetch us the count of Notifications that were unseen by the  User, so that we can display it in the based Template, next to the name of the User.

NotificationController.php => return in a Json Response the count of fetched by the NotificationRepository.